Refactor: pages 폴더에 라우팅 되는 페이지들 추가
