+++ src/router.js
@@ -1,15 +1,15 @@
 import Vue from 'vue';
 import VueRouter from 'vue-router';
-import BoardsComponent from './components/BoardsComponent.vue';
-import Main from './components/Main.vue';
-import Page404 from './components/Page404.vue';
+import Boards from './pages/Boards.vue';
+import Main from './pages/Main.vue';
+import Page404 from './pages/Page404.vue';
 import './style/reset.scss';
 
 Vue.use(VueRouter);
 
 const routes = [
   { path: '/', component: Main },
-  { path: '/boards', component: BoardsComponent },
+  { path: '/boards', component: Boards },
   { path: '*', component: Page404 }
 ];
